keep the commented out texting feature in same component
